[contracts] separate test files, track NFT contract deployer
